Make BART run on the scanner.
- mmapping on TMPDIR /tmp/share does not work
on the scanner, where this is a CIFS mounted
directory.
/tmp should be a tmpfs in the chroot on the scanner,
thus this should not fill up the chroot either.

- ENV variables aren't carried through to the chroot-image.
Thus 'import bart' wouldn't work in the chroot.
(on the github main branch we already have an installable
python module; but for now stay at v0.9.00)

- static linking leads to segfault in gfortran lib.
To resolve this for now, changed to dynamic linking and
install the runtime dependencies in the container.

- Switched bart-build container to same base image as the fire-python
container.